Make the bootstrap command perform a fake bootstrap
The fake-juju cli bootstrap command now posts an HTTP request against
fake-jujud's control plane API in order to perform a fake "bootstrap",
where fake-jujud starts a new controller machine (machine 0).

The logic in the FakeJujuService.InitializeController() method, which
is in charge of implementing the fake bootstrap, has been modified to
not only create the machine (in the "pending" state), but also to
transition it to the "started" state.

The full end-to-end process is exercised by new python integration
test assertions.