Implement the /bootstrap control plane API endpoint
This branch adds a new "/bootstrap" endpoint to the internal control
plane API for the fake-jujud process.

It can be used to perform a fake "bootstrap" of a controller bound to
the dummy provider.

This will happen in the "suite.SetUpTest(c)" call which is basically
what the HTTP API triggers (under the hood the juju testing machinery
will setup the controller).

The branch also adds a minimal HTTP client for talking to the control
plane API and hitting the "/bootstrap" endpoint.

A test has been added to demonstrate that the client and server work
end-to-end. More testing will be added in follow-up branches to assert
that a working controller gets actually greated.